Fix metric

+++ anchor_free/model/metric.py
@@ -47,10 +47,12 @@ class BoundaryIoU(torchmetrics.Metric):
         for i in range(b):
             for j in range(c):
                 boundary_target = mask_to_boundary(
-                    target[i][j].cpu().numpy(), boundary_size=self.boundary_size
+                    target[i][j].cpu().numpy().astype("uint8"),
+                    boundary_size=self.boundary_size,
                 )
                 boundary_preds = mask_to_boundary(
-                    preds[i][j].cpu().numpy(), boundary_size=self.boundary_size
+                    preds[i][j].cpu().numpy().astype("uint8"),
+                    boundary_size=self.boundary_size,
                 )
                 self.iou(boundary_preds, boundary_target)
